Made the AA tree bottom node externally visible
This change allows static initialization of dan_aa_tree
structures, and is absolutely necessary for users to
traverse an AA tree themselves (they need to check for
pointers to the bottom node). It also made the AA node
and AA trees conform better to the STRUCT_INIT system
developing in libdan.